[Symbol] Add DeclVendor::FindTypes
Summary:
Following up on the plan I outlined in D63622, we can remove the
dependence on clang in all the places where we only want to find the
types from the DeclVendor. This means that currently DeclVendor depends
on clang, but centralizing the dependency makes it easier to refactor
cleanly.

Differential Revision: https://reviews.llvm.org/D63853

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@364962 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 843f5f54c0b39567e01ff62e874ff3e4a1f2886c
apple-llvm-split-dir: lldb/