[yaml2elf] - Treat the SHN_UNDEF section as kind of regular section.
We have a logic that adds a few sections implicitly.
Though the SHT_NULL section with section number 0
is an exception.

In D64913 I want to teach yaml2obj to redefine the null section.
And in this patch I add it to the sections list,
to make it kind of a regular section.

Differential revision: https://reviews.llvm.org/D65087

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@366785 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 4c73ef533f0bbc9b8ccc9f7e745037ec01a84bc9)

apple-llvm-split-commit: 54e3c4420d31de965aeebe481a9fb08af6997b7f
apple-llvm-split-dir: llvm/