[ABISysV] Fix regression for Simulator and MacABI
The ABISysV ABI was refactored in r364216 to support the Windows ABI for
x86_64. In particular it changed ABISysV_x86_64::CreateInstance to
switch on the OS type. This breaks debugging MacABI apps as well as apps
in the simulator. This adds back the necessary cases.

We have a test on Github that exercises this code path and which I'd
like to upstream once the remaining MacABI parts become available in
clang.

Differential revision: https://reviews.llvm.org/D67869

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@372642 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 0792da1ad9df81c8b048b472f585f758df1ad08f)

apple-llvm-split-commit: 5a0dc4adcff267295eac50d9690e0efab91133b2
apple-llvm-split-dir: lldb/