[debugserver] Add --version/-V command line option to debugserver.
When not running under a TTY the output is buffered and not flushed
before debugserver exits which makes it impossible to parse the version
string. This adds a -V/--version command that just prints the version to
stdout and exits with an exit code zero.

Differential revision: https://reviews.llvm.org/D68156

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@373127 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 93ee5f460d77ad058b17c309def0678a9bba613c)

apple-llvm-split-commit: 2492881e59d12c1f85dcd985e542e6f5d29faf09
apple-llvm-split-dir: lldb/