build: serialise `LLVM_ENABLE_UNWIND_TABLES` into LLVMConfig
Serialize the value of the configuration option into the configuration so that
builds which integrate LLVM can identify the value of the flag that was used to
build the libraries.  This is intended to be used by Swift to control tests
which rely on the unwind information.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@373253 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit e9915e833b601c0ff32e1a196f37645cb7e61d48)

apple-llvm-split-commit: 50f1e3e500b98a62c6e870aaa5064262304e7732
apple-llvm-split-dir: llvm/